add seven segement, and new template with palin 7d number

+++ panumbaweb/interface/models.py
@@ -7,7 +7,7 @@ from django.db import models
 class AIcontact(models.Model):
     name = models.CharField(max_length=100)
     question = models.TextField()
-    answer = models.IntegerField(default=0)
+    answer = models.BigIntegerField(default=0)
     context = models.TextField()
     
     def __str__(self):
